Move Firestore hooks from package to app layer
- Move React hooks from firestore-generator package to app layer
- Rename hooks with more descriptive names
- Add Japanese documentation comments
- Fix unused imports
- Remove hooks export from firestore-generator package
- Fix build issues
