+++ apps/web/src/hooks/firestore/index.ts
@@ -1,0 +1,3 @@
+export * from './useFirestoreEntity';
+export * from './useFirestoreHistory';
+export * from './useFirestoreCollection';
