+++ packages/firestore-generator/src/index.ts
@@ -6,6 +6,3 @@ export * from './generator';
 // Event sourcing
 export * from './events';
 export * from './repository';
-
-// React hooks
-export * from './hooks';
